+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## OpenRaider (0.1.3) xythobuz <xythobuz@xythobuz.de>
 
+    [ 20141217 ]
+    * Introduced texel-offset in getUV() in an attempt to fix the texture-bleeding
+    * Mesh is now called RoomMesh
+
     [ 20141216 ]
     * Allow navigation with a free-floating Camera
     * Room Meshes are displayed more or less correctly